requirements.txt
guppy3
PySimpleGUI